% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_lindley.R
\name{ttestEvidence}
\alias{ttestEvidence}
\title{Justify your alpha level by avoiding the Lindley paradox or aiming for moderate or strong evidence when using a t-test.}
\usage{
ttestEvidence(evidence, n1, n2 = 0, Cauchy = T)
}
\arguments{
\item{evidence}{Desired level of evidence: "Lindley" to avoid the Lindley Paradox, "moderate" to achieve moderate evidence and "strong" to achieve strong evidence.
Users that are more familiar with Bayesian statistics can also directly enter their desired Bayes factor.}

\item{n1}{Sample size in Group 1.}

\item{n2}{Sample size in Group 2. Leave blank for a one-sample or paired-sample t.test.}

\item{Cauchy}{If true the Cauchy prior with scale 0.707 is used if false, the Unit information Bayes factor is used.}
}
\value{
alpha level required for a two-sample t-test.
}
\description{
Justify your alpha level by avoiding the Lindley paradox or aiming for moderate or strong evidence when using a t-test.
}
\section{References}{

to be added
}

\examples{
## Avoid the Lindley paradox for a two sample t-test with 300 participants per condition
ttestEvidence("lindley", 300, 300)
}
